Wire in XmlDoc2CmdletDoc to generate documentation from source
Wire in @redgate/XmlDoc2CmdletDoc as a post-build step to generate
the PowerShell MAML documentation from XML doc comments in the source.

Currently only `Get-GcsBucket` is documented, but should serve as
a good example for future cmdlets.

The change also updates the Google API client libraries to v1.11.1,
which no longer has the woes related to versioning.