Change Instance cmdlets. (#80)
Change GceInstance cmdlets to be more in line with the newer written cmdlets.

Unify the parameters of Add-GceInstance, New-GceInstanceConfig and Add-GceInstanceTemplate. The now have mostly the same parameters, and use much of the same code to build the request objects. This was accomplished by creating common superclasses.
Make these cmdlets take strongly typed Image and Disk objects.
Add a Get-GceMachineType cmdlet.