Very large commit here, I'm sorry. Anchor links was a more intense process than I expected.
With the prior implementation of the website, anchor links were not possible because of too many parts of the site being reliant on parameters controlled by controllers. To fix this, I had to rewrite a significant amount of the underlying frontend- but not many changes to the HTML were made (beyond importing new files and variables/etc). I also had to migrate to hosting the website on appengine in order to utilize the benefits app.yaml provides.
All files should probably be looked at again.

GSJLint and the google html/css style guide were both used. May be a few small issues I missed though.