Drop code-font on "Cloud Tools for PowerShell"
Especially with three full-fixed-width spaces in the name, putting it in `<code>` looks totally awkward. In any case, `<code>` should be reserved for things you will actually type at the command line, which you never do with "Cloud Tools for PowerShell", unlike `gcloud` or even `Add-GcdManagedZone`.

Also rewords text with an incorrect use of "latter" in the last paragraph. It seems "later" was meant, but even then it was in need of  more extensive restructuring for English grammar and clarity.