Add missing Google Cloud DNS to products.json & home.html (#279)
* Add Google Cloud DNS to products.json



This is needed for it to appear in the left navbar of the Reference tab. I work as part of this team at Google (alexdupuy@).

The Cloud DNS cmdlets page https://googlecloudplatform.github.io/google-cloud-powershell/cmdlets/google-cloud-dns.html gets a 404, although the dns-cmdlets branch was merged. I don't see the file(s) in the latest merge to website from master PR#275 either, so that also needs to be regenerated with the Cloud DNS docs (probably after merging PR#274) but other action may be needed.

* Add Google Cloud DNS example to website home.html

Google Cloud DNS got overlooked here as well.